Move load_true_labels from depth_analysis.py to preprocessing.py and use there in suitable cases.
